requirements.txt
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4.0"
django==5.0.6 ; python_version >= "3.12" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.12" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.12" and python_version < "4.0"
opencv-python==4.9.0.80 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.3.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg-binary==3.1.19 ; implementation_name != "pypy" and python_version >= "3.12" and python_version < "4.0"
psycopg[binary]==3.1.19 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.0 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
